LanguageApi/requirements.txt
fastapi~=0.110.0
uvicorn~=0.29.0
classy-fastapi~=0.6.1
pydantic~=2.6.4
pyjwt~=2.8.0
keybert~=0.8.4
flair~=0.13.1
keyphrase-vectorizers~=0.0.11
starlette~=0.36.3
openai~=1.14.2
boto3~=1.34.67
azure-ai-textanalytics~=5.3.0
yake~=0.4.8
sentence-transformers~=2.5.1
simple-injection~=0.2.5
python-dotenv~=1.0.1